+++ src/TextClient.cpp
@@ -334,7 +334,7 @@ int CamulecmdApp::ProcessCommand(int CmdId)
 						// If the user requested all, then we select all files and exit the loop
 						// since there is little point to add anything more to "everything"
 						if( token == wxT("all") ) {
-							for(int i = 0;i < reply_all->GetTagCount();i++) {
+							for (size_t i = 0; i < reply_all->GetTagCount(); i++) {
 	                                                	const CECTag *tag = reply_all->GetTagByIndex(i);
 	                                                	if (tag) {
 	                                                        	request->AddTag(CECTag(EC_TAG_PARTFILE, tag->GetMD4Data()));
@@ -348,7 +348,7 @@ int CamulecmdApp::ProcessCommand(int CmdId)
 							}
 						} else {
 							 // Go through the dl queue and look at each filename
-							for(int i = 0; i < reply_all->GetTagCount(); i++) {
+							for (size_t i = 0; i < reply_all->GetTagCount(); i++) {
 								CEC_PartFile_Tag *tag = (CEC_PartFile_Tag *)reply_all->GetTagByIndex(i);
 								if (tag) {
 									wxString partmetname = tag->PartMetName();
@@ -608,7 +608,7 @@ wxString StatTree2Text(CEC_StatTree_Node_Tag *tree, int depth)
 		return wxEmptyString;
 	}
 	wxString result = wxString(wxChar(' '), depth) + tree->GetDisplayString() + wxT("\n");
-	for (int i = 0; i < tree->GetTagCount(); ++i) {
+	for (size_t i = 0; i < tree->GetTagCount(); ++i) {
 		CEC_StatTree_Node_Tag *tmp = (CEC_StatTree_Node_Tag*)tree->GetTagByIndex(i);
 		if (tmp->GetTagName() == EC_TAG_STATTREE_NODE) {
 			result += StatTree2Text(tmp, depth + 1);
@@ -675,7 +675,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 			}
 			break;
 		case EC_OP_STRINGS:
-			for (int i = 0; i < response->GetTagCount(); ++i) {
+			for (size_t i = 0; i < response->GetTagCount(); ++i) {
 				const CECTag *tag = response->GetTagByIndex(i);
 				if (tag) {
 					s << tag->GetStringData() << wxT("\n");
@@ -735,7 +735,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 			break;
 		}
 		case EC_OP_DLOAD_QUEUE:
-			for(int i = 0; i < response->GetTagCount(); ++i) {
+			for (size_t i = 0; i < response->GetTagCount(); ++i) {
 				CEC_PartFile_Tag *tag =
 					(CEC_PartFile_Tag *)response->GetTagByIndex(i);
 				if (tag) {
@@ -765,7 +765,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 			}
 			break;
 		case EC_OP_ULOAD_QUEUE:
-			for(int i = 0; i < response->GetTagCount(); ++i) {
+			for (size_t i = 0; i < response->GetTagCount(); ++i) {
 				const CECTag *tag = response->GetTagByIndex(i);
 				const CECTag *clientName = tag ? tag->GetTagByName(EC_TAG_CLIENT_NAME) : NULL;
 				const CECTag *partfileName = tag ? tag->GetTagByName(EC_TAG_PARTFILE_NAME) : NULL;
@@ -782,7 +782,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 			}
 			break;
 		case EC_OP_LOG:
-			for (int i = 0; i < response->GetTagCount(); ++i) {
+			for (size_t i = 0; i < response->GetTagCount(); ++i) {
 				const CECTag *tag = response->GetTagByIndex(i);
 				if (tag) {
 					s << tag->GetStringData() << wxT("\n");
@@ -791,7 +791,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 			}
 			break;
 		case EC_OP_SERVER_LIST:
-			for(int i = 0; i < response->GetTagCount(); i ++) {
+			for (size_t i = 0; i < response->GetTagCount(); i++) {
 				const CECTag *tag = response->GetTagByIndex(i);
 				const CECTag *serverName = tag ? tag->GetTagByName(EC_TAG_SERVER_NAME) : NULL;
 				if (tag && serverName) {
@@ -808,7 +808,7 @@ void CamulecmdApp::Process_Answer_v2(const CECPacket *response)
 		case EC_OP_SEARCH_RESULTS:
 			m_Results_map.clear();
 			s << wxString::Format(_("Number of search results: %i\n"),response->GetTagCount());
-			for (int i = 0;i < response->GetTagCount();i++) {
+			for (size_t i = 0; i < response->GetTagCount(); i++) {
 				CEC_SearchFile_Tag *tag = (CEC_SearchFile_Tag *)response->GetTagByIndex(i);
 				//printf("Tag FileName: %s \n",(const char*)unicode2char(tag->FileName()));
 				//if (tag != NULL)
